// src/app/pages/interfaces/promoline-interface.ts
export interface PromolinePrice {

    Producto: string;
    Precio: number;

}

export interface PromolineActiveItem {
    clave: string;
}

export interface PromolineProductSingle {

    Producto: string;
    Apartado: number;
    Categoria: string;
    SubCategoria: string;
    Color: string;
    Compras: number;
    Precio: number;
    Descripcion: string;
    Disponible: number;
    Externo: number;
    Fecha: string;
    Imagen: string;
    Total: number;

}